Update search.html
It doesn't necessarily have to say this but the way it's currently written suggests you're introducing something new when you're really just finishing the process. I was confused by this and thought the plugin had lost functionality and found others online with the same issue.

Alternately it could say "1 more step! Adding the new search engine via the URL bar"